chore: rename bluesky.description.ts to bluesky-description-converter.ts
