+++ apps/client-server/src/app/websites/implementations/bluesky/bluesky.website.ts
@@ -49,7 +49,7 @@ import { MessageWebsite } from '../../models/website-modifiers/message-website';
 import { OAuthWebsite } from '../../models/website-modifiers/oauth-website';
 import { WithCustomDescriptionParser } from '../../models/website-modifiers/with-custom-description-parser';
 import { Website } from '../../website';
-import { BlueskyConverter } from './bluesky.description';
+import { BlueskyConverter } from './bluesky-description-converter';
 import { BlueskyFileSubmission } from './models/bluesky-file-submission';
 import { BlueskyMessageSubmission } from './models/bluesky-message-submission';
